Changed Camera move method to use floats in -1 to 1 range for PTZ.

+++ src/us/mn/state/dot/tms/comm/manchester/CommandRequest.java
@@ -62,7 +62,7 @@ public class CommandRequest extends Request {
 
 	/** Encode a speed value for pan/tilt command */
 	static byte encodeSpeed(int v) {
-		return (Math.abs(v) - 1) << 1;
+		return (byte)((Math.abs(v) - 1) << 1);
 	}
 
 	/** Encode a pan command packet */
